// src/_services/tg-bot/__generated__/SendMediaGroup.ts
/* eslint-disable */
/* tslint:disable */
// @ts-nocheck
/*
 * ---------------------------------------------------------------
 * ## THIS FILE WAS GENERATED VIA SWAGGER-TYPESCRIPT-API        ##
 * ##                                                           ##
 * ## AUTHOR: acacode                                           ##
 * ## SOURCE: https://github.com/acacode/swagger-typescript-api ##
 * ---------------------------------------------------------------
 */

import {
  SendMediaGroupCreateData,
  SendMediaGroupCreateError,
  SendMediaGroupCreatePayload,
} from "./data-contracts";
import { ContentType, HttpClient, RequestParams } from "./http-client";

export class SendMediaGroup<SecurityDataType = unknown> {
  http: HttpClient<SecurityDataType>;

  constructor(http: HttpClient<SecurityDataType>) {
    this.http = http;
  }

  /**
   * @description Use this method to send a group of photos, videos, documents or audios as an album. Documents and audio files can be only grouped in an album with messages of the same type. On success, an array of [Messages](https://core.telegram.org/bots/api/#message) that were sent is returned.
   *
   * @name SendMediaGroupCreate
   * @request POST:/sendMediaGroup
   */
  sendMediaGroupCreate = (
    data: SendMediaGroupCreatePayload,
    params: RequestParams = {},
  ) =>
    this.http.request<SendMediaGroupCreateData, SendMediaGroupCreateError>({
      path: `/sendMediaGroup`,
      method: "POST",
      body: data,
      type: ContentType.FormData,
      format: "json",
      ...params,
    });
}
